requirements.txt
ollama>=0.4.0
yfinance>=0.2.50
pandas>=2.0.0
feedparser>=6.0.11
requests>=2.31.0
rich>=13.7.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
